Fixed cert enrollment problem with empty rangeUnit in profile.
Previously cert enrollment might fail after editing the profile
using the console. This is because the console added an empty
rangeUnit parameter, but the server rejected the empty value.

The convertRangeUnit() methods in several classes have been
modified to accept the empty value and convert it into the
default value (i.e. day).

https://fedorahosted.org/pki/ticket/2308