Fix old KRA servlets to check realm
The old KRA servlets to list and display keys do not go through
the same code paths as the REST API.  Therefore, they do not
check the authz realm.

This patch adds the relevant code.  No attempt is made to fix the
error handling of the old servlets.  the long term solution for this
is to deprecate the old servlets and make the UI use the REST API
instead.  Therefore, authz failures due to realm checks are handled
in the same way as other authz changes.