Fixed TPS VLV sort orders.
The TPS VLVs for tokens and activities has been modified to sort
the results by date in reverse order.

The DBRegistry.getLDAPAttributes() was modified to support reverse
sort order by recognizing the "-" prefix in the list of sort keys
and pass it to LDAP.

The DBVirtualList.setSortKey() was modified to ignore bubble up the
exceptions that happen during LDAP attribute mapping.

https://fedorahosted.org/pki/ticket/2263
https://fedorahosted.org/pki/ticket/2269