Fixed pki-server subsystem-cert-update.
The pki-server subsystem-cert-update is supposed to restore the
system certificate data and requests into CS.cfg. The command was
broken since the CASubsystem class that contains the code to find
the certificate requests from database was not loaded correctly.
To fix the problem the CASubsystem class has been moved into the
pki/server/__init__.py.

All pki-server subsystem-* commands have been modified to check
the validity of the instance.

An option has been added to the pki-server subsystem-cert-show
command to display the data and request of a particular system
certificate.

The redundant output of the pki-server subsystem-cert-update has
been removed. The updated certificate data and request can be
obtained using the pki-server subsystem-cert-show command.

https://fedorahosted.org/pki/ticket/2385