Updated RPM spec.
The code in the RPM spec that moves the upgrade scripts has been
updated to reflect the FixSELinuxContexts deletion.

The libselinux-python is used by deployment and upgrade scripts
to set the SELinux contexts, so a direct runtime dependency has
been added to the RPM spec file.

The duplicate python-ldap and python-lxml dependencies have been
removed.

https://fedorahosted.org/pki/ticket/2452