Fixed TPS UI system menu.
The TPS UI has been modified to adjust the system menu based
on the list of accessible components obtained during login.

The TPSApplication has been modified to use TPSAccountService
which returns the list of accessible components based on the
following properties in the CS.cfg:
* admin: target.configure.list
* agent: target.agent_approve.list

The AccountInfo has been changed to extend the ResourceMessage
such that it can be used to pass the list of accessible
components as an attribute.

https://fedorahosted.org/pki/ticket/2523
(cherry picked from commit f979c3b436e9a12e8c71ba0abab5c892d375f945)