Modify retrieval and archival mechanisms in KRA REST
When clients call retrieveKey(), three possible alternatives
now obtain:
1. client passes in an approved request. Request is processed
   and the secret is retrieved.
2. client passes in key_id and wrapping parameters and either:
  a) request can be processed immediately and synchronously
     and request is created, and secret is returned.
  b) request cannot be processed immediately.  Recovery request
     is created and request_id returned to the client

Depending on server configuration, the requests in case (2a)
will be stored in ldap or will be ephemeral (in memory only).

More complicated realm based logic to determine if requests
can be processed synchronously or ephemerally will be added in
a later patch.