Add groups and request attributes to external principals
Add the ExternalAuthenticationValve valve, which, if an externally
authenticated principal is available, reads the REMOTE_USER_GROUP
information from the Coyote request and adds the groups ("roles" in
Tomcat terminology) to the principal.

It also saves a complete copy of the request attribute map in the
princpial.  The new class ExternalPrincipal is used to achieve this.

Part of: https://pagure.io/dogtagpki/issue/1359