Add IAuthToken implementation for external principals
Many parts of Dogtag expect an IAuthToken, which represents the
authenticated user.  The sole implementation, AuthToken, uses some
concepts that do not carry across to externally authenticated
principals, e.g. an external principal does not have an associated
IAuthManager that was used to authenticate the principal.  Therefore
something different is needed.

Implement ExternalAuthToken which wraps a GenericPrincipal and
provides access to the data therein.

Part of: https://pagure.io/dogtagpki/issue/1359