Resolve: pkispawn does not change default ecc key size from nistp256 when nistp384 is specified in spawn config
Ticket #2552.

This fix turned out simple. The client was correctly setting the required data, but it was putting the curveName in the
"keySize" field of the SystemCertData object sent to the back end. The configuration routine was trying to find the name in the "curveName" field when its really in the "keySize" field. This issue is restricted to the ECC case. It is fine to simply fix this in the server, since the "keySize" is a string anyway and it makes decent sense.

(cherry picked from commit ae350a3d4e0ae9b82fa44ebdfa37654f0083b4c1)