X500Name: add method to get all attributes of a given type
To implement a profile default that copies the CN to a SAN dNSName,
we need to examine the CN values present in the Subject DN.
Specifically, we want to look at the "most specific" CN value.  The
'getCommonName' method returns the "least specific" value in the
name, thus is not suitable.

Add the 'getAttributesForOid(ObjectIdentifier)' method, which
returns an ordered list of values of the given name attribute type,
from least specific to most specific.

Part of: https://fedorahosted.org/pki/ticket/1710