Add profile component that copies CN to SAN dNSName
Add the 'CommonNameToSANDefault' profile default class.  When used
on a profile, this will examine the (most-specific) Common Name in
the Subject DN, and if it looks like a DNS name, will add it to the
Subject Alternative Name extension, creating the extension if it
does not already exist.

Also add upgrade scriptlet to add the component to registry.cfg in
existing installations.

Fixes: https://fedorahosted.org/pki/ticket/1710