Allow arbitrary user data in cert request
If a certificate request comes with additional data in the
'cert-request' query param, add that to the request.  Profile
components can then use this data.

This is needed to convey the subject principal name to the
ExternalProcessConstraint, when validating FreeIPA certificate
requests after we switch to GSS-API authentication.

Part of: https://pagure.io/dogtagpki/issue/1359