ISourceConfigStore: add clear() method to interface
The SourceConfigStore load() method does not clear the config store,
but this might be necessary to avoid stale data if wanting to
perform a complete replacement of the data (e.g. reload from file).

We should not change the behaviour of load() in case some code is
relying on the current behaviour, so add the clear() method to the
interface.

Part of: https://fedorahosted.org/pki/ticket/2588

Change-Id: Ia139a49f1a23c4f9410d7b94c9a4c8f14f29fe93