CertProcessor: set external principal attributes into request
When processing a certificate request, if the authenticated
principal is an ExternalPrincipal, add its whole attribute map to
the IRequest.  This provides a way for AJP request attributes to be
propagated through the profile system to profile components like
ExternalProcessConstraint.  One such attribute that is needed for
GSS-API support is "KRB5CCNAME".

Part of: https://pagure.io/dogtagpki/issue/1359